An implicit variable was not found for an operation. This was fixed.
Also a strange name generation was fixed:
For an expression such as:

transactions->select(oclIsTypeOf(Earning)).points

an implicit 'collect' operation is generated, in that form:

transactions->select(oclIsTypeOf(Earning))->collect(points)

For the 'collect' operation a new variable must be generated. In the version before that bug fix, the new variable had the name of the source expression, in this case 'select'. Now the variable has a name that reflects the generated origin.